refactor(55-01): extract runtime, events, publish modules from fuse lib.rs
Task 1 of plan 55-01. Move NETWORK_TIMEOUT/block_with_timeout into runtime.rs;
pending/event types + spawn_metadata_refresh into events.rs; publish queue/
coordinator + next_file_publish_sequence + resolve helpers into publish.rs.
lib.rs keeps cfg-gated module decls + re-exports so the public crate surface is
unchanged. cargo build + cargo test -p cipherbox-fuse pass (64/64); winfsp
feature build is Windows-only (CI-gated).
